Fixed the calculation issue

+++ src/main/java/io/sskuratov/sodiumconsumptioncalc/calculations/Formula.java
@@ -1,8 +1,6 @@
 package io.sskuratov.sodiumconsumptioncalc.calculations;
 
-import java.math.BigDecimal;
-
 public interface Formula {
 
-    BigDecimal evaluate();
+    Double evaluate();
 }
