+++ src/main/java/io/sskuratov/sodiumconsumptioncalc/constraints/Weight.java
@@ -2,11 +2,9 @@ package io.sskuratov.sodiumconsumptioncalc.constraints;
 
 import io.sskuratov.sodiumconsumptioncalc.exceptions.InputException;
 
-import java.math.BigDecimal;
-
-public class Weight implements Constraint<BigDecimal> {
+public class Weight implements Constraint<Double> {
     @Override
-    public void validate(BigDecimal value) throws InputException {
+    public void validate(Double value) throws InputException {
         if ((value.compareTo(getMax()) > 0) || (value.compareTo(getMin()) < 0)) {
 
             String builder = "Допустимые значения для данного параметра находятся в интервале от " +
@@ -21,12 +19,12 @@ public class Weight implements Constraint<BigDecimal> {
         }
     }
 
-    private BigDecimal getMax() {
-        return BigDecimal.valueOf(300L);
+    private Double getMax() {
+        return 300.0;
     }
 
-    private BigDecimal getMin() {
-        return BigDecimal.valueOf(20L);
+    private Double getMin() {
+        return 20.0;
     }
 
     @Override
